requirements.txt
python-telegram-bot==20.7
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3
pytesseract==0.3.10
python-dotenv==1.0.0